src/requirements.txt
Flask==2.0.0
Jinja2==3.0.0
python-dotenv==0.17.1